+++ RevenueCatUI/CustomerCenter/Views/ManageSubscriptionsView.swift
@@ -1,4 +1,12 @@
 //
+//  Copyright RevenueCat Inc. All Rights Reserved.
+//
+//  Licensed under the MIT License (the "License");
+//  you may not use this file except in compliance with the License.
+//  You may obtain a copy of the License at
+//
+//      https://opensource.org/licenses/MIT
+//
 //  ManageSubscriptionsView.swift
 //
 //
@@ -161,8 +169,8 @@ struct ManageSubscriptionsView_Previews: PreviewProvider {
 
     static var previews: some View {
         let viewModel = ManageSubscriptionsViewModel(
-            configuration: CustomerCenterTestData.customerCenterData,
-            subscriptionInformation: CustomerCenterTestData.subscriptionInformation)
+            configuration: CustomerCenterConfigTestData.customerCenterData,
+            subscriptionInformation: CustomerCenterConfigTestData.subscriptionInformation)
         ManageSubscriptionsView(viewModel: viewModel)
     }
