+++ RevenueCatUI/CustomerCenter/Views/NoSubscriptionsView.swift
@@ -1,4 +1,12 @@
 //
+//  Copyright RevenueCat Inc. All Rights Reserved.
+//
+//  Licensed under the MIT License (the "License");
+//  you may not use this file except in compliance with the License.
+//  You may obtain a copy of the License at
+//
+//      https://opensource.org/licenses/MIT
+//
 //  NoSubscriptionsView.swift
 //
 //
